Renaming SourceHandlers classes
Update readme
