+++ readme.md
@@ -1,5 +1,8 @@
-Po Parser
-=========
+PoParser
+=======
+PoParser is a personal project to fulfill a need I got: parse Gettext Portable files (*.po files) and edit its content using PHP.  
+
+PoParser requires PHP >= 5.4, but may work in 5.3 too.
 
 [![Latest Stable Version](https://poser.pugx.org/sepia/po-parser/v/stable)](https://packagist.org/packages/sepia/po-parser) 
 [![Total Downloads](https://poser.pugx.org/sepia/po-parser/downloads)](https://packagist.org/packages/sepia/po-parser) 
@@ -7,13 +10,12 @@ Po Parser
 [![Build Status](https://travis-ci.org/raulferras/PHP-po-parser.png?branch=master)](https://travis-ci.org/raulferras/PHP-po-parser) 
 [![Code Coverage](https://scrutinizer-ci.com/g/raulferras/PHP-po-parser/badges/coverage.png?s=a19ece2a8543b085ab1a5db319ded3bc4530b567)](https://scrutinizer-ci.com/g/raulferras/PHP-po-parser/) 
 [![Scrutinizer Quality Score](https://scrutinizer-ci.com/g/raulferras/PHP-po-parser/badges/quality-score.png?s=6aaf3c31ce15cebd1d4bed718cd41fd2d921fd31)](https://scrutinizer-ci.com/g/raulferras/PHP-po-parser/) 
-[![Gitter](https://badges.gitter.im/Join Chat.svg)](https://gitter.im/raulferras/PHP-po-parser?utm_source=badge&utm_medium=badge&utm_campaign=pr-badge&utm_content=badge)
+
+[![Gitter](https://badges.gitter.im/raulferras/PHP-po-parser.svg)](https://gitter.im/raulferras/PHP-po-parser?utm_source=badge&utm_medium=badge&utm_campaign=pr-badge)
 
 
-PoParser is a personal project to fulfill a need I got: parse Gettext Portable files (*.po files) and edit its content using PHP.  
-
-PoParser will allow you to read PO Data from any source (files and strings built-in), update it and store back to a file (or get the compiled string).
-
+Features
+========
 It supports following parsing features:
 
 - header section.
@@ -25,71 +27,56 @@ It supports following parsing features:
 - # keys (translator comments).
 - #. keys (Comments extracted from source code).
 - #: keys (references).
-- #| keys (previously untranslated), both single and multiline.
+- #| keys (previously strings), both single and multiline.
 - #~ keys (old entries), both single and multiline.
+
+Installation
+============
+
+```
+composer require sepia/po-parser
+```
 
 Usage
 =====
+```
+<?php 
+// Parse a po file
+$fileHandler = new Sepia\FileSystem('es.po');
 
-    // Parse a po file
-    $fileHandler = new Sepia\FileHandler('es.po');
-    
-    $poParser = new Sepia\PoParser($fileHandler);
-    $entries  = $poParser->parse();
-    // $entries contains every entry in es.po file.
+$poParser = new Sepia\Parser($fileHandler);
+$catalog  = $poParser->parse();
 
-    // Update entries
-    $msgid = 'Press this button to save';
-    $entries[$msgid]['msgstr'] = 'Pulsa este botón para guardar';
-    $poParser->setEntry($msgid, $entries[$msgid]);
-    // You can also change translator comments, code comments, flags...
+// Get an entry
+$entry = $catalog->getEntry('welcome.user');
 
+// Update entry
+$entry = new Entry('welcome.user', 'Welcome User!');
+$catalog->setEntry($entry);
 
+// You can also modify other entry attributes as translator comments, code comments, flags...
+$entry->setTranslatorComments(array('This is shown whenever a new user registers in the website'));
+$entry->setFlags(array('fuzzy', 'php-code'));
+```
 
-Changelog
-=========
-v5.0 (WIP)
-* Classes are now fluid.
-* Namespaces reorganized.
-* Removed `fuzzy` index in favour of `flags`.
-* Display line number on parsing errors instead of line content.
-* Adds compatibility with `#~|` entries.
-* `parseString()` and `parseFile()` converted to factory methods.
-* Removed method `updateEntry()` in favour of `setEntry()`.
-
-v4.2.2
-* More PHPDocs fixes
-* Strict comparisons used where safe.
-* Fix example for `writeFile`.
-* Support for EOL line formatting.
-
-v4.2.1
-* Support multiline for plural entries (thanks @Ben-Ho)
-
-v4.2.0
-* Add function to add plural and context to existing entry (thanks @Ben-Ho)
-* Add ability to change msg id of entry (thanks @wildex)
-
-
-v4.1.1
-* Fixes with multi-flags entries (thanks @gnouet)
-
-v4.1
-* Constructor now accepts options to define separator used in multiline msgid entries.
-* New method `getOptions()`.
-
-v4.0
-
-* new methods parseString() and parseFile() replace the old parse()`
-* new method writeFile() replaces the old write().
-* new method compile() which takes all parsed entries and coverts back to a PO formatted string.
-
-[See whole changelog](https://github.com/raulferras/PHP-po-parser/wiki/Changelog)
-
+## Save Changes back to a file
+Use `PoCompiler` to together with `FileSystem` to save a catalog back to a file:
+ 
+```
+$fileHandler = new FileSystem('en.po');
+$compiler = new PoCompiler();
+$fileHandler->save(
+    $compiler->compile($catalog),
+    $this->resourcesPath.'temp.po'
+);
+```
 
 Documentation
 =============
-[See v4 documentation](https://github.com/raulferras/PHP-po-parser/wiki/Documentation-4.0)
+[v5 Documentation](https://github.com/raulferras/PHP-po-parser/wiki/Documentation-5.0)
+[v4 documentation](https://github.com/raulferras/PHP-po-parser/wiki/Documentation-4.0)
+
+[Migration guide from v4 to v5](https://github.com/raulferras/PHP-po-parser/wiki/Migration-v4-to-v5)
 
 
 Testing
@@ -100,14 +87,3 @@ To execute tests, from command line type:
 ```
 php vendor/bin/phpunit
 ```
-
-
-Install via composer
-====================
-Edit your composer.json file to include the following:
-
-    {
-        "require": {
-            "sepia/po-parser": "dev-master"
-        }
-    }
